Move atomic parsing functions to name.py
The following functions were moved: atomic_symbol, isotope_symbol,
element_name, atomic_number, _extract_charge_state, _is_neutron,
_is_hydrogen, _is_electron, _is_positron, and _is_antiproton.
